Move utils in separate folders
